+++ lib/flutter_ocr_native.dart
@@ -1,5 +1,8 @@
 export 'src/models/ocr_exception.dart';
 export 'src/models/ocr_result.dart';
+export 'src/models/ocr_watermark.dart';
 export 'src/ocr_reader.dart';
 export 'src/ocr_platform_interface.dart';
+export 'src/utils/ocr_document_saver.dart';
 export 'src/validators/ocr_validator.dart';
+export 'src/widgets/ocr_document_viewer.dart';